move revision stuff out of clean and into it's own helper
